requirements.txt
glfw
PyOpenGL
PyOpenGL-Accelerate
numpy
pillow